# internal/handler/user.go
package handler

import (
	"net/http"

	"go-blog-api/internal/db"
	"go-blog-api/internal/model"
	"go-blog-api/internal/utils"

	"github.com/gin-gonic/gin"
	"golang.org/x/crypto/bcrypt"
)

type RegisterRequest struct {
	Username string `json:"username" binding:"required,min=3,max=20"`
	Password string `json:"password" binding:"required,min=6,max=20"`
}

// Register 处理用户注册请求
// @Summary 用户注册
// @Description 注册新用户（账号需3-20位，密码需6-20位)
// @Tags 用户模块
// @Accept json
// @Produce json
// @Param request body RegisterRequest true "账号和密码"
// @Success 200 {object} map[string]interface{} "注册成功"
// @Failure 400 {object} map[string]interface{} "参数错误，账号需3-20位，密码需6-20位"
// @Failure 409 {object} map[string]interface{} "用户名已存在"
// @Failure 500 {object} map[string]interface{} "密码加密失败"
// @Failure 500 {object} map[string]interface{} "用户创建失败"
// @Router /register [post]
func Register(c *gin.Context) {
	var req RegisterRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "参数错误（账号需3-20位，密码需6-20位）"})
		return
	}
	// 检查用户名是否已存在
	var count int64
	db.DB.Model(&model.User{}).Where("username = ?", req.Username).Count(&count)
	if count > 0 {
		c.JSON(http.StatusBadRequest, gin.H{"error": "用户名已存在"})
		return
	}
	// 密码加密
	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(req.Password), bcrypt.DefaultCost)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "密码加密失败"})
		return
	}
	user := model.User{
		Username:       req.Username,
		PasswordDigest: string(hashedPassword),
	}
	if err := db.DB.Create(&user).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "用户创建失败"})
		return
	}
	c.JSON(http.StatusOK, gin.H{
		"message": "注册成功",
		"data": gin.H{
			"id":       user.ID,
			"username": user.Username,
		},
	})
}

type LoginRequest struct {
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required"`
}

// Login 处理用户登录请求
// @Summary 用户登录
// @Description 校验账号密码并返回 JWT Token
// @Tags 用户模块
// @Accept json
// @Produce json
// @Param request body LoginRequest true "账号和密码"
// @Success 200 {object} map[string]interface{} "登录成功，返回 JWT Token"
// @Failure 400 {object} map[string]interface{} "参数错误（账号和密码不能为空）"
// @Failure 401 {object} map[string]interface{} "用户名或密码错误"
// @Failure 500 {object} map[string]interface{} "系统异常，生成 Token 失败"
// @Router /login [post]
func Login(c *gin.Context) {
	var req LoginRequest

	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "参数错误，请提供账号和密码"})
		return
	}

	var user model.User
	// 根据用户名查询用户
	if err := db.DB.Where("username = ?", req.Username).First(&user).Error; err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "用户名或密码错误"})
		return
	}
	// 密码校验
	if err := bcrypt.CompareHashAndPassword([]byte(user.PasswordDigest), []byte(req.Password)); err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "用户名或密码错误"})
		return
	}

	// 生成JWT Token
	token, err := utils.GenerateToken(user.ID, user.Username)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "系统异常，生成Token失败"})
		return
	}

	//把Token返回给前端
	c.JSON(http.StatusOK, gin.H{
		"message": "登录成功",
		"token":   token,
	})
}